refactor: 미션 그룹 관련 패키지 및 클래스 구조 변경 미션 그룹 관련 클래스와 패키지 이름을 변경하여 일관성을 높였습니다.
