+++ src/main/java/org/example/hugmeexp/domain/missionGroup/repository/MissionGroupRepository.java
@@ -1,6 +1,6 @@
-package org.example.hugmeexp.domain.mission.repository;
+package org.example.hugmeexp.domain.missionGroup.repository;
 
-import org.example.hugmeexp.domain.mission.entity.MissionGroup;
+import org.example.hugmeexp.domain.missionGroup.entity.MissionGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
